Fix: go/txteditor/buffer.NextLine> +1 is not required.
```
This is test text
2nd line is here
```

when buf.point is on 1st line, 'eolpoint' is 17(newline after "text").
Suppose we are at first 't' of "test":

```
This is test text
        ^-------- buf.point
2nd line is here
```

Then `c.x` will be 12. And by calling `buffer.NextLine` it is expected
to move point to whitespace between "line" and "is":

```
This is test text
2nd line is here
        ^-------- new buf.point
```

As '2' of "2nd" is at point eol+1, we simply need to add c.x to eol